+++ src/app/components/sidebar/sidebar.component.ts
@@ -28,6 +28,8 @@ import { environment } from '../../../environments/environment';
 // import brand from 'assets/brand/brand.json';
 import { BrandService } from '../../services/brand.service';
 import { WsRequestsService } from './../../services/websocket/ws-requests.service';
+import { WsMsgsService } from './../../services/websocket/ws-msgs.service';
+
 import { LoggerService } from './../../services/logger/logger.service';
 import { avatarPlaceholder, getColorBck } from '../../utils/util'
 declare const $: any;
@@ -182,6 +184,9 @@ export class SidebarComponent implements OnInit, AfterViewInit {
     CONTACT_EDIT_ROUTE_IS_ACTIVE: boolean;
     CONTACT_CONVS_ROUTE_IS_ACTIVE: boolean;
 
+    IS_REQUEST_FOR_PANEL_ROUTE: boolean;
+    IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE: boolean;
+
     prjct_profile_name: string;
 
     prjct_trial_expired: boolean;
@@ -212,7 +217,13 @@ export class SidebarComponent implements OnInit, AfterViewInit {
     prjct_name: string;
     private unsubscribe$: Subject<any> = new Subject<any>();
     current_selected_prjct: any;
+    new_messages_count: number;
 
+    NOTIFICATION_SOUND: string;
+    storedValuePrefix = 'dshbrd----'
+    hasPlayed = false
+    currentUrl: string;
+    audio: any;
     constructor(
         private router: Router,
         public location: Location,
@@ -229,7 +240,8 @@ export class SidebarComponent implements OnInit, AfterViewInit {
         private deptService: DepartmentService,
         public brandService: BrandService,
         public wsRequestsService: WsRequestsService,
-        private logger: LoggerService
+        private logger: LoggerService,
+        private wsMsgsService: WsMsgsService
     ) {
         this.logger.log('[SIDEBAR] !!!!! HELLO SIDEBAR')
 
@@ -248,14 +260,11 @@ export class SidebarComponent implements OnInit, AfterViewInit {
         this.translateChangeAvailabilitySuccessMsg();
         this.translateChangeAvailabilityErrorMsg();
         this.getProfileImageStorage();
-
-        
-
         this.getUserAvailability();
         this.getUserUserIsBusy();
         this.getProjectUserId();
 
-        
+
 
         this.hasChangedAvailabilityStatusInUsersComp();
 
@@ -273,7 +282,90 @@ export class SidebarComponent implements OnInit, AfterViewInit {
         this.getChatUrl();
         this.isMac();
         this.listenHasDeleteUserProfileImage();
+        this.listenToForegroundNotificationCount();
+        this.listenSoundPreference();
+        this.getNotificationSoundPreferences();
     }
+
+
+
+
+    getNotificationSoundPreferences() {
+        // NOTIFICATION_SOUND = 'enabled';
+        const storedNotificationSound = localStorage.getItem(this.storedValuePrefix + 'sound');
+
+        if (storedNotificationSound !== 'undefined' && storedNotificationSound !== null) {
+
+            this.NOTIFICATION_SOUND = storedNotificationSound;
+            console.log('[SIDEBAR] NOTIFICATION_SOUND -  this.NOTIFICATION_SOUND', this.NOTIFICATION_SOUND)
+        } else {
+            this.NOTIFICATION_SOUND = 'enabled';
+            console.log('[SIDEBAR] NOTIFICATION_SOUND -  this.NOTIFICATION_SOUND', this.NOTIFICATION_SOUND)
+        }
+    }
+
+    listenSoundPreference() {
+        this.wsRequestsService.hasChangedSoundPreference$
+            .pipe(
+                takeUntil(this.unsubscribe$)
+            )
+            .subscribe((newSoundPreference) => {
+                console.log('[SIDEBAR] - LISTEN TO SOUND PREFERNCE CHANGED ', newSoundPreference);
+                this.NOTIFICATION_SOUND = newSoundPreference;
+            }, error => {
+                console.error('[SIDEBAR] - LISTEN TO SOUND PREFERNCE CHANGED * ERROR * ', error)
+            }, () => {
+                console.log('[SIDEBAR] - LISTEN TO SOUND PREFERNCE CHANGED *** COMPLETE *** ')
+            });
+    }
+
+    listenToForegroundNotificationCount() {
+        this.wsRequestsService.foregroundNotificationCount$
+            .pipe(
+                takeUntil(this.unsubscribe$)
+            )
+            .subscribe((foregroundNoticationCount) => {
+                console.log('[SIDEBAR] - FOREGROUND NOTIFICATION COUNT ', foregroundNoticationCount);
+                this.new_messages_count = foregroundNoticationCount;
+
+                if (this.NOTIFICATION_SOUND === 'enabled' && this.IS_REQUEST_FOR_PANEL_ROUTE === false && this.IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE === false) {
+                    // this.logger.log('[NAVBAR] NOTIFICATION_SOUND (showNotification) hasPlayed ', this.hasPlayed)
+                    if (this.hasPlayed === false) {
+                        // this.logger.log('[NAVBAR] NOTIFICATION_SOUND (showNotification) hasPlayed (HERE IN IF)', this.hasPlayed)
+                        this.audio = new Audio();
+
+                        this.audio.src = 'assets/pling.mp3';
+
+                        this.audio.load();
+
+                        this.audio.play().then(() => {
+
+                            this.hasPlayed = true
+                            console.log('[SIDEBAR] - SOUND HAS PLAYED  hasPlayed ', this.hasPlayed)
+
+                            setTimeout(() => {
+                                this.hasPlayed = false;
+                                console.log('[SIDEBAR] - SOUND HAS PLAYED  hasPlayed ', this.hasPlayed)
+
+                            }, 4000);
+                        }).catch((error: any) => {
+                            console.log('[APP-COMP] ***soundMessage error*', error);
+                        });
+
+                    }
+
+                }
+
+
+
+            }, error => {
+                console.error('[SIDEBAR] - FOREGROUND NOTIFICATION COUNT * ERROR * ', error)
+            }, () => {
+                console.log('[SIDEBAR] - FOREGROUND NOTIFICATION COUNT *** COMPLETE *** ')
+            });
+    }
+
+
 
     getLoggedUser() {
         this.auth.user_bs.subscribe((user) => {
@@ -487,16 +579,30 @@ export class SidebarComponent implements OnInit, AfterViewInit {
         if (!this.public_Key.includes("PAY")) {
             this.isVisiblePAY = false;
         }
-
     }
 
 
     getCurrentRoute() {
         this.router.events.filter((event: any) => event instanceof NavigationEnd)
             .subscribe(event => {
+                if (event.url.indexOf('/request-for-panel') !== -1) {
+                    this.IS_REQUEST_FOR_PANEL_ROUTE = true;
+                    console.log('[NAVBAR] NavigationEnd - IS_REQUEST_FOR_PANEL_ROUTE  ', this.IS_REQUEST_FOR_PANEL_ROUTE);
+                } else {
+                    this.IS_REQUEST_FOR_PANEL_ROUTE = false;
+                    console.log('[NAVBAR] NavigationEnd - IS_REQUEST_FOR_PANEL_ROUTE  ', this.IS_REQUEST_FOR_PANEL_ROUTE);
+                }
+
+                if (event.url.indexOf('/unserved-request-for-panel') !== -1) {
+                    this.IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE = true;
+                    console.log('[SIDEBAR] NavigationEnd - IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE  ', this.IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE);
+                } else {
+                    this.IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE = false;
+                    console.log('[SIDEBAR] NavigationEnd- IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE  ', this.IS_UNSERVEDREQUEST_FOR_PANEL_ROUTE);
+                }
 
                 if (event.url.indexOf('/autologin') !== -1) {
-                    this.logger.log('[SIDEBAR] NavigationEnd - THE activities-demo route IS ACTIVE  ', event.url);
+                    console.log('[SIDEBAR] NavigationEnd - THE activities-demo route IS ACTIVE  ', event.url);
                     this.AUTOLOGIN_ROUTE_IS_ACTIVE = true;
 
                 } else {
@@ -787,13 +893,6 @@ export class SidebarComponent implements OnInit, AfterViewInit {
                     this.CONTACT_CONVS_ROUTE_IS_ACTIVE = false;
                     // console.log('[SIDEBAR] NavigationEnd - CONTACT_CONVS_ROUTE_IS_ACTIVE ', this.CONTACT_CONVS_ROUTE_IS_ACTIVE);
                 }
-
-
-
-
-
-
-
             });
     }
 
@@ -1098,6 +1197,8 @@ export class SidebarComponent implements OnInit, AfterViewInit {
 
 
     }
+
+
     // NO MORE USED - SUBSTITUDED WITH changeAvailabilityState
     // availale_unavailable_status(hasClickedChangeStatus: boolean) {
     //     hasClickedChangeStatus = hasClickedChangeStatus;
@@ -1127,7 +1228,7 @@ export class SidebarComponent implements OnInit, AfterViewInit {
             if (this.project) {
 
                 // this.getDeptsAndFilterDefaultDept();
-                
+
                 this.projectId = this.project._id
                 this.getProjectUserRole(this.projectId);
                 this.findCurrentProjectAmongAll(this.projectId)