+++ src/app/services/users-local-db.service.ts
@@ -126,6 +126,16 @@ export class LocalDbService {
     return isWide
   }
 
+  storeForegrondNotificationsCount(count) : void {
+    localStorage.setItem(this.prefix + 'foregroundcount', count);
+    console.log('HEY - SAVE IN STORAGE !!! Foregrond Notifications Count ', count);
+  }
+
+  getForegrondNotificationsCount() {
+    const foregroundcCunt = localStorage.getItem(this.prefix + 'foregroundcount');
+    return foregroundcCunt
+  }
+
 
 }